Refactored pill and projectile controllers to be more modular

+++ client/Assets/Scripts/AbilityEffects/DamageEffect.cs
@@ -9,7 +9,7 @@ namespace pillz.client.Scripts.AbilityEffects
         public override void Execute(uint playerId, Rigidbody2D target, in ExplosionHit hit)
         {
             // Scale damage by proximity (at least 1)
-            var dmg = (uint)Mathf.Max(1, Mathf.RoundToInt(MaxDamage * hit.Falloff));
+            var dmg = Mathf.Max(1, Mathf.RoundToInt(MaxDamage * hit.Falloff));
             GameHandler.Connection.Reducers.ApplyDamage(playerId, dmg);
         }
     }
